Nach mehreren Minuten (ca 20) auftretender Abbruch der Foregroundservices und Entfernung der Notification wurde (nach Ergebnissen erster Tests) beseitigt. 10 Sekunden Inaktivität Trigger der Übung funktioniert. Der 20min Test, Hintergrundaktivität, dann Ablehnung, Caancellation des Monitoring und dann manueller Start und wieder Monitoring-Start haben in dieser Reihenfolge funktioniert! Nachrichten von Smartphone kommen weiterhin an

+++ health-services/ExerciseSampleCompose/app/src/main/java/com/example/healthapp/service/ExerciseServiceMonitor.kt
@@ -67,7 +67,7 @@ constructor(
     @SuppressLint("RestrictedApi")
     private fun processExerciseUpdate(exerciseUpdate: ExerciseUpdate) {
         // Dismiss any ongoing activity notification.
-        if (exerciseUpdate.exerciseStateInfo.state.isEnded) {
+        if (exerciseUpdate.exerciseStateInfo.state.isEnded && exerciseClientManager.monitoringEnded.value) {
             exerciseService.removeOngoingActivityNotification()
         }
